requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.5
pybit==5.11.0
python-dotenv==1.1.1
psutil==7.0.0
aiohttp==3.12.15
python-binance==1.0.29
pydantic==2.11.7
pydantic-settings==2.10.1
PyYAML==6.0.2